% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kao.R
\name{kao_random}
\alias{kao_random}
\title{Return a random kaomoji}
\usage{
kao_random(x = list(kaomoji1, kaomoji2))
}
\arguments{
\item{x}{A list of kaomoji sets to sample from}
}
\description{
Return a random kaomoji
}
