% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaomoji-package.R
\docType{package}
\name{kaomoji-package}
\alias{kaomoji}
\alias{kaomoji-package}
\title{kaomoji: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
