% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kaomoji1}
\alias{kaomoji1}
\alias{kaomoji2}
\title{Kaomojis}
\source{
\url{https://github.com/brnnkt/kaomoji_list}

\url{http://kaomoji.ru/en}
}
\usage{
kaomoji1

kaomoji2
}
\description{
Various kaomojis, each set comes from a different source.
}
\keyword{datasets}
